Set Useragent for connectors. Extract settings classes to their own module
