+++ java/extensions/semantickernel-settings-loader/src/main/java/com/microsoft/semantickernel/connectors/ai/openai/util/OpenAIClientProvider.java
@@ -7,7 +7,10 @@ import com.azure.ai.openai.OpenAIAsyncClient;
 import com.azure.ai.openai.OpenAIClientBuilder;
 import com.azure.core.credential.AzureKeyCredential;
 import com.azure.core.credential.KeyCredential;
+import com.azure.core.util.ClientOptions;
+import com.microsoft.semantickernel.SemanticKernelHttpSettings;
 import com.microsoft.semantickernel.exceptions.ConfigurationException;
+import com.microsoft.semantickernel.settings.SettingsMap;
 import java.io.File;
 import java.util.Collections;
 import java.util.List;
@@ -24,6 +27,7 @@ import org.slf4j.Logger;
  * OpenAI client.
  */
 public class OpenAIClientProvider {
+
     private static final Logger LOGGER =
             org.slf4j.LoggerFactory.getLogger(OpenAIClientProvider.class);
 
@@ -45,6 +49,7 @@ public class OpenAIClientProvider {
 
     @Nullable private final ClientType clientType;
     private final Map<String, String> configuredSettings;
+    private final ClientOptions clientOptions;
 
     /**
      * Create a new OpenAI client provider
@@ -56,6 +61,7 @@ public class OpenAIClientProvider {
             Map<String, String> configuredSettings, @Nullable ClientType clientType) {
         this.configuredSettings = Collections.unmodifiableMap(configuredSettings);
         this.clientType = clientType;
+        this.clientOptions = SemanticKernelHttpSettings.getUserAgent(configuredSettings);
     }
 
     private static OpenAIClientProvider buildProvider(
@@ -222,6 +228,7 @@ public class OpenAIClientProvider {
 
         return new OpenAIClientBuilder()
                 .credential(new KeyCredential(settings.getKey()))
+                .clientOptions(clientOptions)
                 .buildAsyncClient();
     }
 
@@ -239,6 +246,7 @@ public class OpenAIClientProvider {
         OpenAIClientBuilder builder =
                 new OpenAIClientBuilder()
                         .endpoint(settings.getEndpoint())
+                        .clientOptions(clientOptions)
                         .credential(new AzureKeyCredential(settings.getKey()));
 
         return builder.buildAsyncClient();
