+++ java/pom.xml
@@ -73,6 +73,7 @@
 
         <module>extensions/semantickernel-planners</module>
         <module>extensions/semantickernel-gpt3-tokenizer</module>
+        <module>extensions/semantickernel-settings-loader</module>
 
         <module>plugins/semantickernel-plugin-core</module>
