+++ java/semantickernel-bom/pom.xml
@@ -95,6 +95,11 @@
                 <artifactId>semantickernel-connectors-memory-azurecognitivesearch</artifactId>
                 <version>${project.version}</version>
             </dependency>
+            <dependency>
+                <groupId>com.microsoft.semantic-kernel</groupId>
+                <artifactId>semantickernel-settings-loader</artifactId>
+                <version>${project.version}</version>
+            </dependency>
 
             <dependency>
                 <groupId>com.azure</groupId>
